Set up the molecule cluster UT env.
1. For the operator deployment,
a. Use the operator.yml file directly to test the real world
deployment.
b. I've also enhanced the template operator.yaml.j2 file so
we can still choose to use that if that's desired somehow.

2. Create the Contract CRD in the prepare phase. Also add
the Contract testing in the verify phase.

4. For the destroy task,
a. The destroy sequence should do both CRDs first so that
operator can clean up the APIC resources accordingly.
b. Also add the operator itself to the destroy list.
c. The namespace deletion should be the last one to go
otherwise it will automatically delete all the resources
under this namespace. Also we don't really need to delete
the namespace here coz we have to use aci-containers-system
namespace for this testing.

5. Populate the namespace and operator_image parameters in the
molecule.yml file.