Make the molecule test-local scenario work
Need to deploy the aci-controller-config and aci-user-cert first
before the operator container can be up and running. We need
these 2 yaml files to reference some APIC in order to deploy
those EPG and contract related resources thru the operator.
Here I'm using the config from the hyperflex k8s21 cluster for
this purpose.